avoid implicit conversions

+++ ThrEd4/utf8conv.h
@@ -31,7 +31,7 @@ inline auto utf8ToUtf16(std::string const& utf8) -> std::wstring {
   auto const inLength  = gsl::narrow_cast<int>(utf8.size());
   auto const outLength = MultiByteToWideChar(CP_UTF8, 0, utf8.data(), inLength, nullptr, 0);
 
-  std::wstring utf16(outLength, '\0');
+  std::wstring utf16(gsl::narrow_cast<size_t>(outLength), '\0');
   MultiByteToWideChar(CP_UTF8, 0, utf8.data(), inLength, utf16.data(), outLength);
   return utf16;
 }
@@ -40,7 +40,7 @@ inline auto utf16ToUtf8(std::wstring const& utf16) -> std::string {
   auto const inLength  = gsl::narrow_cast<int>(utf16.size());
   auto const outLength = WideCharToMultiByte(CP_UTF8, 0, utf16.data(), inLength, nullptr, 0, nullptr, nullptr);
 
-  std::string utf8(outLength, '\0');
+  std::string utf8(gsl::narrow_cast<size_t>(outLength), '\0');
   WideCharToMultiByte(CP_UTF8, 0, utf16.data(), inLength, utf8.data(), outLength, nullptr, nullptr);
   return utf8;
 }
